+++ src/afterglow/controllers/launchpad_pro.clj
@@ -468,8 +468,7 @@
             (with-show (:show controller)
               (if (and active (not (:held cue)))
                 (show/end-effect! (:key cue))
-                (let [vars (controllers/starting-vars-for-velocity cue velocity)
-                      id (show/add-effect-from-cue-grid! cue-x cue-y :var-overrides vars)
+                (let [id (show/add-effect-from-cue-grid! cue-x cue-y :velocity velocity)
                       holding (and (:held cue) (not @(:shift-mode controller)))]
                   (controllers/add-overlay
                    (:overlays controller)